Start cosmetic refactor #2263
