+++ randomizer/Patching/ApplyLocal.py
@@ -15,12 +15,12 @@ import js
 from randomizer.Enums.Models import Model, ModelNames, HeadResizeImmune
 from randomizer.Enums.Settings import RandomModels, BigHeadMode
 from randomizer.Lists.Songs import ExcludedSongsSelector
+from randomizer.Patching.Cosmetics.TextRando import writeCrownNames
+from randomizer.Patching.Cosmetics.Holiday import applyHolidayMode
 from randomizer.Patching.CosmeticColors import (
     apply_cosmetic_colors,
-    applyHolidayMode,
     overwrite_object_colors,
     writeMiscCosmeticChanges,
-    writeCrownNames,
     darkenDPad,
     darkenPauseBubble,
 )
