+++ randomizer/Patching/Cosmetics/ModelSwaps.py
@@ -1,0 +1,333 @@
+from randomizer.Enums.Models import Model, Sprite
+
+turtle_models = [
+    Model.Diddy,  # Diddy
+    Model.DK,  # DK
+    Model.Lanky,  # Lanky
+    Model.Tiny,  # Tiny
+    Model.Chunky,  # Regular Chunky
+    Model.ChunkyDisco,  # Disco Chunky
+    Model.Cranky,  # Cranky
+    Model.Funky,  # Funky
+    Model.Candy,  # Candy
+    Model.Seal,  # Seal
+    Model.Enguarde,  # Enguarde
+    Model.BeaverBlue_LowPoly,  # Beaver
+    Model.Squawks_28,  # Squawks
+    Model.KlaptrapGreen,  # Klaptrap Green
+    Model.KlaptrapPurple,  # Klaptrap Purple
+    Model.KlaptrapRed,  # Klaptrap Red
+    Model.KlaptrapTeeth,  # Klaptrap Teeth
+    Model.SirDomino,  # Sir Domino
+    Model.MrDice_41,  # Mr Dice
+    Model.Beetle,  # Beetle
+    Model.NintendoLogo,  # N64 Logo
+    Model.MechanicalFish,  # Mech Fish
+    Model.ToyCar,  # Toy Car
+    Model.BananaFairy,  # Fairy
+    Model.Shuri,  # Starfish
+    Model.Gimpfish,  # Gimpfish
+    Model.Spider,  # Spider
+    Model.Rabbit,  # Rabbit
+    Model.KRoolCutscene,  # K Rool
+    Model.SkeletonHead,  # Skeleton Head
+    Model.Vulture_76,  # Vulture
+    Model.Vulture_77,  # Racing Vulture
+    Model.Tomato,  # Tomato
+    Model.Fly,  # Fly
+    Model.SpotlightFish,  # Spotlight Fish
+    Model.Puftup,  # Pufftup
+    Model.CuckooBird,  # Cuckoo Bird
+    Model.IceTomato,  # Ice Tomato
+    Model.Boombox,  # Boombox
+    Model.KRoolFight,  # K Rool (Boxing)
+    Model.Microphone,  # Microbuffer
+    Model.DeskKRool,  # K Rool's Desk
+    Model.Bell,  # Bell
+    Model.BonusBarrel,  # Bonus Barrel
+    Model.HunkyChunkyBarrel,  # HC Barrel
+    Model.MiniMonkeyBarrel,  # MM Barrel
+    Model.TNTBarrel,  # TNT Barrel
+    Model.Rocketbarrel,  # RB Barrel
+    Model.StrongKongBarrel,  # SK Barrel
+    Model.OrangstandSprintBarrel,  # OSS Barrel
+    Model.BBBSlot_143,  # BBB Slot
+    Model.PlayerCar,  # Tiny Car
+    Model.Boulder,  # Boulder
+    Model.Boat_158,  # Boat
+    Model.Potion,  # Potion
+    Model.ArmyDilloMissle,  # AD Missile
+    Model.TagBarrel,  # Tag Barrel
+    Model.QuestionMark,  # Question Mark
+    Model.Krusha,  # Krusha
+    Model.BananaPeel,  # Banana Peel
+    Model.Butterfly,  # Butterfly
+    Model.FunkyGun,  # Funky's Gun
+]
+
+panic_models = [
+    Model.Diddy,  # Diddy
+    Model.DK,  # DK
+    Model.Lanky,  # Lanky
+    Model.Tiny,  # Tiny
+    Model.Chunky,  # Regular Chunky
+    Model.ChunkyDisco,  # Disco Chunky
+    Model.Cranky,  # Cranky
+    Model.Funky,  # Funky
+    Model.Candy,  # Candy
+    Model.Seal,  # Seal
+    Model.Enguarde,  # Enguarde
+    Model.BeaverBlue_LowPoly,  # Beaver
+    Model.Squawks_28,  # Squawks
+    Model.KlaptrapGreen,  # Klaptrap Green
+    Model.KlaptrapPurple,  # Klaptrap Purple
+    Model.KlaptrapRed,  # Klaptrap Red
+    Model.MadJack,  # Mad Jack
+    Model.Troff,  # Troff
+    Model.SirDomino,  # Sir Domino
+    Model.MrDice_41,  # Mr Dice
+    Model.RoboKremling,  # Robo Kremling
+    Model.Scoff,  # Scoff
+    Model.Beetle,  # Beetle
+    Model.NintendoLogo,  # N64 Logo
+    Model.MechanicalFish,  # Mech Fish
+    Model.ToyCar,  # Toy Car
+    Model.Klump,  # Klump
+    Model.Dogadon,  # Dogadon
+    Model.BananaFairy,  # Fairy
+    Model.Guard,  # Guard
+    Model.Shuri,  # Starfish
+    Model.Gimpfish,  # Gimpfish
+    Model.KLumsy,  # K Lumsy
+    Model.Spider,  # Spider
+    Model.Rabbit,  # Rabbit
+    # Model.Beanstalk,  # Beanstalk
+    Model.KRoolCutscene,  # K Rool
+    Model.SkeletonHead,  # Skeleton Head
+    Model.Vulture_76,  # Vulture
+    Model.Vulture_77,  # Racing Vulture
+    Model.Ghost,  # Ghost
+    Model.Fly,  # Fly
+    Model.FlySwatter_83,  # Fly Swatter
+    Model.Owl,  # Owl
+    Model.Book,  # Book
+    Model.SpotlightFish,  # Spotlight Fish
+    Model.Puftup,  # Pufftup
+    Model.Mermaid,  # Mermaid
+    Model.Mushroom,  # Mushroom Man
+    Model.Worm,  # Worm
+    Model.EscapeShip,  # Escape Ship
+    Model.KRoolFight,  # K Rool (Boxing)
+    Model.Microphone,  # Microbuffer
+    Model.BonusBarrel,  # Bonus Barrel
+    Model.HunkyChunkyBarrel,  # HC Barrel
+    Model.MiniMonkeyBarrel,  # MM Barrel
+    Model.TNTBarrel,  # TNT Barrel
+    Model.Rocketbarrel,  # RB Barrel
+    Model.StrongKongBarrel,  # SK Barrel
+    Model.OrangstandSprintBarrel,  # OSS Barrel
+    Model.PlayerCar,  # Tiny Car
+    Model.Boulder,  # Boulder
+    Model.VaseCircle,  # Vase
+    Model.VaseColon,  # Vase
+    Model.VaseTriangle,  # Vase
+    Model.VasePlus,  # Vase
+    Model.ArmyDilloMissle,  # AD Missile
+    Model.TagBarrel,  # Tag Barrel
+    Model.QuestionMark,  # Question Mark
+    Model.Krusha,  # Krusha
+    Model.Light,  # Light
+    Model.BananaPeel,  # Banana Peel
+    Model.FunkyGun,  # Funky's Gun
+]
+
+bother_models = [
+    Model.BeaverBlue_LowPoly,  # Beaver
+    Model.Klobber,  # Klobber
+    Model.Kaboom,  # Kaboom
+    Model.KlaptrapGreen,  # Green Klap
+    Model.KlaptrapPurple,  # Purple Klap
+    Model.KlaptrapRed,  # Red Klap
+    Model.KlaptrapTeeth,  # Klap Teeth
+    Model.Krash,  # Krash
+    Model.Troff,  # Troff
+    Model.NintendoLogo,  # N64 Logo
+    Model.MechanicalFish,  # Mech Fish
+    Model.Krossbones,  # Krossbones
+    Model.Rabbit,  # Rabbit
+    Model.SkeletonHead,  # Minecart Skeleton Head
+    Model.Tomato,  # Tomato
+    Model.IceTomato,  # Ice Tomato
+    Model.GoldenBanana_104,  # Golden Banana
+    Model.Microphone,  # Microbuffer
+    Model.Bell,  # Bell
+    Model.Missile,  # Missile (Car Race)
+    Model.Buoy,  # Red Buoy
+    Model.BuoyGreen,  # Green Buoy
+    Model.RarewareLogo,  # Rareware Logo
+]
+
+piano_models = [
+    Model.Krash,
+    Model.RoboKremling,
+    Model.KoshKremling,
+    Model.KoshKremlingRed,
+    Model.Kasplat,
+    Model.Guard,
+    Model.Krossbones,
+    Model.Mermaid,
+    Model.Mushroom,
+    Model.GoldenBanana_104,
+    Model.FlySwatter_83,
+    Model.Ruler,
+]
+piano_extreme_model = [
+    Model.SkeletonHead,
+    Model.Owl,
+    Model.Kosha,
+    # Model.Beanstalk,
+]
+
+spotlight_fish_models = [
+    # Model.Turtle,  # Lighting Bug
+    Model.Seal,
+    Model.BeaverBlue,
+    Model.BeaverGold,
+    Model.Zinger,
+    Model.Squawks_28,
+    Model.Klobber,
+    Model.Kaboom,
+    Model.KlaptrapGreen,
+    Model.KlaptrapPurple,
+    Model.KlaptrapRed,
+    Model.Krash,
+    # Model.SirDomino,  # Lighting issue
+    # Model.MrDice_41,  # Lighting issue
+    # Model.Ruler, # Lighting issue
+    # Model.RoboKremling, # Lighting issue
+    Model.NintendoLogo,
+    Model.MechanicalFish,
+    Model.ToyCar,
+    Model.Kasplat,
+    Model.BananaFairy,
+    Model.Guard,
+    Model.Gimpfish,
+    # Model.Shuri,  # Lighting issue
+    Model.Spider,
+    Model.Rabbit,
+    Model.KRoolCutscene,
+    Model.KRoolFight,
+    # Model.SkeletonHead, # Lighting bug
+    # Model.Vulture_76, # Lighting bug
+    # Model.Vulture_77, # Lighting bug
+    # Model.Bat, # Lighting bug
+    # Model.Tomato, # Lighting bug
+    # Model.IceTomato, # Lighting bug
+    # Model.FlySwatter_83, # Lighting bug
+    Model.SpotlightFish,
+    Model.Microphone,
+    # Model.Rocketbarrel,  # Model too big, obstructs view
+    # Model.StrongKongBarrel,  # Model too big, obstructs view
+    # Model.OrangstandSprintBarrel,  # Model too big, obstructs view
+    # Model.MiniMonkeyBarrel,  # Model too big, obstructs view
+    # Model.HunkyChunkyBarrel,  # Model too big, obstructs view
+]
+candy_cutscene_models = [
+    Model.Cranky,
+    # Model.Funky, # Disappears with collision
+    Model.Candy,
+    Model.Snide,
+    Model.Seal,
+    Model.BeaverBlue,
+    Model.BeaverGold,
+    Model.Klobber,
+    Model.Kaboom,
+    Model.Krash,
+    Model.Troff,
+    Model.Scoff,
+    Model.RoboKremling,
+    Model.Beetle,
+    Model.MrDice_41,
+    Model.MrDice_56,
+    Model.BananaFairy,
+    Model.Rabbit,
+    Model.KRoolCutscene,
+    Model.KRoolFight,
+    Model.Vulture_76,
+    Model.Vulture_77,
+    Model.Tomato,
+    Model.IceTomato,
+    Model.FlySwatter_83,
+    Model.Microphone,
+    Model.StrongKongBarrel,
+    Model.Rocketbarrel,
+    Model.OrangstandSprintBarrel,
+    Model.MiniMonkeyBarrel,
+    Model.HunkyChunkyBarrel,
+    Model.RambiCrate,
+    Model.EnguardeCrate,
+    Model.Boulder,
+    Model.SteelKeg,
+    Model.GoldenBanana_104,
+]
+
+funky_cutscene_models = [
+    Model.Cranky,
+    Model.Candy,
+    Model.Funky,
+    Model.Troff,
+    Model.Scoff,
+    Model.Ruler,
+    Model.RoboKremling,
+    Model.KRoolCutscene,
+    Model.KRoolFight,
+    Model.Microphone,
+]
+
+# Not holding gun
+funky_cutscene_models_extreme = [
+    Model.BeaverBlue,
+    Model.BeaverGold,
+    Model.Klobber,
+    Model.Kaboom,
+    Model.SirDomino,
+    Model.MechanicalFish,
+    Model.BananaFairy,
+    Model.SkeletonHand,
+    Model.IceTomato,
+    Model.Tomato,
+]
+
+boot_cutscene_models = [
+    Model.Turtle,
+    Model.Enguarde,
+    Model.BeaverBlue,
+    Model.BeaverGold,
+    Model.Zinger,
+    Model.Squawks_28,
+    Model.KlaptrapGreen,
+    Model.KlaptrapPurple,
+    Model.KlaptrapRed,
+    Model.BananaFairy,
+    Model.Spider,
+    Model.Bat,
+    Model.KRoolGlove,
+]
+
+melon_random_sprites = [
+    Sprite.BouncingMelon,
+    Sprite.BouncingOrange,
+    Sprite.Coconut,
+    Sprite.Peanut,
+    Sprite.Grape,
+    Sprite.Feather,
+    Sprite.Pineapple,
+    Sprite.CrystalCoconut0,
+    Sprite.DKCoin,
+    Sprite.DiddyCoin,
+    Sprite.LankyCoin,
+    Sprite.TinyCoin,
+    Sprite.ChunkyCoin,
+    Sprite.Fairy,
+    Sprite.RaceCoin,
+]
